Arabic-first onboarding, cash box history, tank transfers, and fixes
- Onboarding wizard: fully translated, defaults to Arabic with a
  language selector (station admins are overwhelmingly Arabic-speaking
  and this is their first screen, before they could pick a language).
- Fixed formatDate/formatDateTime/formatMoney/formatUsd rendering with
  Arabic-Indic digits when the browser's own locale is Arabic — they
  now force en-US like formatNumber already did, so numbers stay
  consistent app-wide regardless of language.
- Bumped the base font size in RTL, since the app's font has no Arabic
  glyphs and the system fallback renders visibly smaller.
- Profit margin percentage precision: 2 -> 4 decimal places.
- Pump counter reading values are now integers (validation, casts, and
  inputs), default to the tank last used on that pump, are grouped by
  fuel type with a per-fuel-type "liters sold today" total, and the
  history table is newest-first with a time column.
- Cash Box: new history section listing every transaction/settlement
  that affected the register (fuel deliveries, including Sadcop's,
  never touch cash so they're excluded); moved the date-range picker
  between the Today and Selected period sections.
- New tank-to-tank fuel transfer feature (same fuel type only), with
  its own history list on the Inventory page.

+++ app/Http/Controllers/OnboardingController.php
@@ -117,7 +117,7 @@ class OnboardingController extends Controller
             'readings' => ['required', 'array'],
             'readings.*.pump_id' => ['required', 'exists:fuel_pumps,id'],
             'readings.*.tank_id' => ['required', 'exists:tanks,id'],
-            'readings.*.reading_value' => ['required', 'numeric', 'min:0'],
+            'readings.*.reading_value' => ['required', 'integer', 'min:0'],
         ]);
 
         foreach ($data['readings'] as $reading) {
